+++ activerecord/test/cases/connection_adapters/connection_handlers_sharding_db_test.rb
@@ -222,6 +222,84 @@ module ActiveRecord
           ENV["RAILS_ENV"] = previous_env
         end
 
+        class SecondaryBase < ActiveRecord::Base
+          self.abstract_class = true
+        end
+
+        def test_retrieves_proper_connection_with_nested_connected_to_on_abstract_classes
+          previous_env, ENV["RAILS_ENV"] = ENV["RAILS_ENV"], "default_env"
+
+          config = {
+            "default_env" => {
+              "primary"  => { "adapter" => "sqlite3", "database" => "test/db/primary.sqlite3" },
+              "primary_replica"  => { "adapter" => "sqlite3", "database" => "test/db/primary.sqlite3", "replica" => true },
+              "primary_shard_one" => { "adapter" => "sqlite3", "database" => "test/db/primary_shard_one.sqlite3" },
+              "primary_shard_one_replica" => { "adapter" => "sqlite3", "database" => "test/db/primary_shard_one.sqlite3", "replica" => true },
+              "primary_shard_two" => { "adapter" => "sqlite3", "database" => "test/db/primary_shard_two.sqlite3" },
+              "primary_shard_two_replica" => { "adapter" => "sqlite3", "database" => "test/db/primary_shard_two.sqlite3", "replica" => true },
+              "secondary"  => { "adapter" => "sqlite3", "database" => "test/db/secondary.sqlite3" },
+              "secondary_replica"  => { "adapter" => "sqlite3", "database" => "test/db/secondary.sqlite3", "replica" => true },
+              "secondary_shard_one" => { "adapter" => "sqlite3", "database" => "test/db/secondary_shard_one.sqlite3" },
+              "secondary_shard_one_replica" => { "adapter" => "sqlite3", "database" => "test/db/secondary_shard_one.sqlite3", "replica" => true },
+              "secondary_shard_two" => { "adapter" => "sqlite3", "database" => "test/db/secondary_shard_two.sqlite3" },
+              "secondary_shard_two_replica" => { "adapter" => "sqlite3", "database" => "test/db/secondary_shard_two.sqlite3", "replica" => true }
+            }
+          }
+
+          @prev_configs, ActiveRecord::Base.configurations = ActiveRecord::Base.configurations, config
+
+          ActiveRecord::Base.connects_to(shards: {
+            default: { writing: :primary, reading: :primary_replica },
+            shard_one: { writing: :primary_shard_one, reading: :primary_shard_one_replica }
+          })
+
+          SecondaryBase.connects_to(shards: {
+            default: { writing: :secondary, reading: :secondary_replica },
+            shard_one: { writing: :secondary_shard_one, reading: :secondary_shard_one_replica },
+            shard_two: { writing: :secondary_shard_two, reading: :secondary_shard_two_replica }
+          })
+
+          assert_equal "primary", ActiveRecord::Base.connection_pool.db_config.name
+          assert_equal "secondary", SecondaryBase.connection_pool.db_config.name
+
+          ActiveRecord::Base.connected_to(role: :reading, shard: :shard_one) do
+            # ActiveRecord:Base is a global switch for passed values
+            assert_equal "primary_shard_one_replica", ActiveRecord::Base.connection_pool.db_config.name
+            assert_equal "secondary_shard_one_replica", SecondaryBase.connection_pool.db_config.name
+
+            SecondaryBase.connected_to(role: :writing, shard: :default) do
+              # Note: Currently there is only granular shard swapping, not role swapping
+              # so this switches the shard only for SecondaryBase but role for both
+              assert_equal "primary_shard_one", ActiveRecord::Base.connection_pool.db_config.name
+              assert_equal "secondary", SecondaryBase.connection_pool.db_config.name
+
+              ActiveRecord::Base.connected_to(role: :writing, shard: :default) do
+                assert_equal "primary", ActiveRecord::Base.connection_pool.db_config.name
+                assert_equal "secondary", SecondaryBase.connection_pool.db_config.name
+              end
+
+              SecondaryBase.connected_to(role: :reading, shard: :shard_two) do
+                # granular switching means only SecondaryBase uses shard_two
+                assert_equal "primary_shard_one_replica", ActiveRecord::Base.connection_pool.db_config.name
+                assert_equal "secondary_shard_two_replica", SecondaryBase.connection_pool.db_config.name
+              end
+
+              ActiveRecord::Base.connected_to(role: :writing) do
+                # shard is inherited for AR::Base from outer most block :shard_one
+                assert_equal "primary_shard_one", ActiveRecord::Base.connection_pool.db_config.name
+                # shard is inherited for SecondaryBase by second block :default
+                assert_equal "secondary", SecondaryBase.connection_pool.db_config.name
+              end
+            end
+          end
+          assert_equal "primary", ActiveRecord::Base.connection_pool.db_config.name
+          assert_equal "secondary", SecondaryBase.connection_pool.db_config.name
+        ensure
+          ActiveRecord::Base.configurations = @prev_configs
+          ActiveRecord::Base.establish_connection(:arunit)
+          ENV["RAILS_ENV"] = previous_env
+        end
+
         def test_connected_to_raises_without_a_shard_or_role
           error = assert_raises(ArgumentError) do
             ActiveRecord::Base.connected_to { }
@@ -322,7 +400,7 @@ module ActiveRecord
         assert ShardConnectionTestModel.find_by_shard_key("foo")
       end
 
-      def test_swapping_shards_in_a_multi_threaded_environment
+      def test_swapping_shards_globally_in_a_multi_threaded_environment
         tf_default = Tempfile.open "shard_key_default"
         tf_shard_one = Tempfile.open "shard_key_one"
 
@@ -417,6 +495,88 @@ module ActiveRecord
         tf_shard_one_reading.unlink
         tf_default_reading.close
         tf_default_reading.unlink
+      end
+
+      def test_swapping_granular_shards_and_roles_in_a_multi_threaded_environment
+        tf_default = Tempfile.open "shard_key_default"
+        tf_shard_one = Tempfile.open "shard_key_one"
+        tf_default_reading = Tempfile.open "shard_key_default_reading"
+        tf_shard_one_reading = Tempfile.open "shard_key_one_reading"
+        tf_default2 = Tempfile.open "shard_key_default2"
+        tf_shard_one2 = Tempfile.open "shard_key_one2"
+        tf_default_reading2 = Tempfile.open "shard_key_default_reading2"
+        tf_shard_one_reading2 = Tempfile.open "shard_key_one_reading2"
+
+        SecondaryBase.connects_to shards: {
+          default: { writing: { database: tf_default.path, adapter: "sqlite3" }, secondary: { database: tf_default_reading.path, adapter: "sqlite3" } },
+          one: { writing: { database: tf_shard_one.path, adapter: "sqlite3" }, secondary: { database: tf_shard_one_reading.path, adapter: "sqlite3" } }
+        }
+
+        SomeOtherBase.connects_to shards: {
+          default: { writing: { database: tf_default2.path, adapter: "sqlite3" }, secondary: { database: tf_default_reading2.path, adapter: "sqlite3" } },
+          one: { writing: { database: tf_shard_one2.path, adapter: "sqlite3" }, secondary: { database: tf_shard_one_reading2.path, adapter: "sqlite3" } }
+        }
+
+        [:default, :one].each do |shard_name|
+          ActiveRecord::Base.connected_to(role: :writing, shard: shard_name) do
+            ShardConnectionTestModel.connection.execute("CREATE TABLE `shard_connection_test_models` (shard_key VARCHAR (255))")
+            ShardConnectionTestModel.connection.execute("INSERT INTO `shard_connection_test_models` VALUES ('shard_key_#{shard_name}')")
+            ShardConnectionTestModelB.connection.execute("CREATE TABLE `shard_connection_test_models` (shard_key VARCHAR (255))")
+            ShardConnectionTestModelB.connection.execute("INSERT INTO `shard_connection_test_models` VALUES ('shard_key_#{shard_name}_b')")
+          end
+
+          ActiveRecord::Base.connected_to(role: :secondary, shard: shard_name) do
+            ShardConnectionTestModel.connection.execute("CREATE TABLE `shard_connection_test_models` (shard_key VARCHAR (255))")
+            ShardConnectionTestModel.connection.execute("INSERT INTO `shard_connection_test_models` VALUES ('shard_key_#{shard_name}_secondary')")
+            ShardConnectionTestModelB.connection.execute("CREATE TABLE `shard_connection_test_models` (shard_key VARCHAR (255))")
+            ShardConnectionTestModelB.connection.execute("INSERT INTO `shard_connection_test_models` VALUES ('shard_key_#{shard_name}_secondary_b')")
+          end
+        end
+
+        shard_one_latch = Concurrent::CountDownLatch.new
+        shard_default_latch = Concurrent::CountDownLatch.new
+
+        ShardConnectionTestModel.connection
+
+        thread = Thread.new do
+          ShardConnectionTestModel.connection
+
+          shard_default_latch.wait
+          assert_equal "shard_key_default", ShardConnectionTestModel.connection.select_value("SELECT shard_key from shard_connection_test_models")
+          assert_equal "shard_key_default_b", ShardConnectionTestModelB.connection.select_value("SELECT shard_key from shard_connection_test_models")
+          shard_one_latch.count_down
+        end
+
+        SecondaryBase.connected_to(shard: :one, role: :secondary) do
+          shard_default_latch.count_down
+          assert_equal "shard_key_one_secondary", ShardConnectionTestModel.connection.select_value("SELECT shard_key from shard_connection_test_models")
+          assert_equal "shard_key_default_secondary_b", ShardConnectionTestModelB.connection.select_value("SELECT shard_key from shard_connection_test_models")
+
+          SomeOtherBase.connected_to(shard: :one, role: :secondary) do
+            assert_equal "shard_key_one_secondary", ShardConnectionTestModel.connection.select_value("SELECT shard_key from shard_connection_test_models")
+            assert_equal "shard_key_one_secondary_b", ShardConnectionTestModelB.connection.select_value("SELECT shard_key from shard_connection_test_models")
+          end
+          shard_one_latch.wait
+        end
+
+        thread.join
+      ensure
+        tf_shard_one.close
+        tf_shard_one.unlink
+        tf_default.close
+        tf_default.unlink
+        tf_shard_one_reading.close
+        tf_shard_one_reading.unlink
+        tf_default_reading.close
+        tf_default_reading.unlink
+        tf_shard_one2.close
+        tf_shard_one2.unlink
+        tf_default2.close
+        tf_default2.unlink
+        tf_shard_one_reading2.close
+        tf_shard_one_reading2.unlink
+        tf_default_reading2.close
+        tf_default_reading2.unlink
       end
     end
   end